Organized log record creation into it's own dir
